Add files via upload
This project was to test our knowledge of the OO Analysis And Design module and how we could analyse and design a large application into smaller building blocks which has to be accomplished in order to not over complicate when coding the application. This was a group project therefore, myself and my colleague had to break down an application for the hope music academy where we were given a complex description of whats actually going inside the academy. Therefore, we had to analyse the description carefully and break it down into smaller pieces before we started to design the application.